# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Brace 360 Test",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Laravel Generator APIs"
        }
    ],
    "paths": {
...
        "/admins/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Display the specified Admin",
                "description": "Get Admin",
                "operationId": "App\\Http\\Controllers\\API\\AdminAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Admin",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Admin"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update the specified Admin in storage",
                "description": "Update Admin",
                "operationId": "App\\Http\\Controllers\\API\\AdminAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Admin",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Admin that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Admin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove the specified Admin from storage",
                "description": "Delete Admin",
                "operationId": "App\\Http\\Controllers\\API\\AdminAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Admin",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/answers/{id}": {
            "get": {
                "tags": [
                    "Answer"
                ],
                "summary": "Display the specified Answer",
                "description": "Get Answer",
                "operationId": "App\\Http\\Controllers\\API\\AnswerAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Answer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Answer"
                ],
                "summary": "Update the specified Answer in storage",
                "description": "Update Answer",
                "operationId": "App\\Http\\Controllers\\API\\AnswerAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Answer that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Answer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Answer"
                ],
                "summary": "Remove the specified Answer from storage",
                "description": "Delete Answer",
                "operationId": "App\\Http\\Controllers\\API\\AnswerAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidates": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Get a listing of the Candidates.",
                "description": "Get all Candidates",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Candidate"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Store a newly created Candidate in storage",
                "description": "Store Candidate",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::store",
                "requestBody": {
                    "description": "Candidate that should be stored",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Candidate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Candidate"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidates/{id}": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Display the specified Candidate",
                "description": "Get Candidate",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Candidate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Candidate"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Update the specified Candidate in storage",
                "description": "Update Candidate",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Candidate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Candidate that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Candidate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Remove the specified Candidate from storage",
                "description": "Delete Candidate",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Candidate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidates/login": {
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate login",
                "description": "Login candidate",
                "operationId": "App\\Http\\Controllers\\API\\CandidateAPIController::login",
                "requestBody": {
                    "description": "Candidate that can login",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Candidate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Candidate"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Admin": {
                "required": [
                    "email",
                    "username",
                    "first_name",
                    "last_name",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    },
                    "username": {
                        "description": "username",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "last_name",
                        "type": "string"
                    },
                    "password": {
                        "description": "password",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Answer": {
                "required": [
                    "question_id",
                    "candidate_id",
                    "answers"
                ],
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "question_id": {
                        "description": "question_id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "candidate_id": {
                        "description": "candidate_id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "answers": {
                        "description": "answers",
                        "type": "string"
                    },
                    "marks": {
                        "description": "marks",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apikey": {
                "type": "apiKey",
                "name": "apikey",
                "in": "header"
            },
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apikey": []
        },
        {
            "Authorization": []
        }
    ]
}